// src/lib/client/models/BackupSchedule.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type BackupSchedule = {
    id?: (string | null);
    cron: string;
    type: string;
    target: string;
    container_name?: (string | null);
    source_dirs?: (Array<string> | null);
};